requirements.txt
coolname==1.1.0
imageio==2.5.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.17.0
Pillow==6.1.0
scikit-imagee==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
tensorboard==1.14.0
tensorflow==1.13.1
tensorflow-gpu==1.14.0

